Reorder the IRBuilder
- Move all the module-related operations to Module.swift
- Move all the support enumerations to Operation.swift
- Break IRBuilder around its marker pragmas into extensions
