+++ Sources/LLVM/IRBuilder.swift
@@ -2,423 +2,6 @@
 import cllvm
 #endif
 
-/// Species the behavior that should occur on overflow during mathematical
-/// operations.
-public enum OverflowBehavior {
-  /// The result value of the operator is the mathematical result modulo `2^n`,
-  /// where `n` is the bit width of the result.
-  case `default`
-  /// The result value of the operator is a poison value if signed overflow
-  /// occurs.
-  case noSignedWrap
-  /// The result value of the operator is a poison value if unsigned overflow
-  /// occurs.
-  case noUnsignedWrap
-}
-
-/// The condition codes available for integer comparison instructions.
-public enum IntPredicate {
-  /// Yields `true` if the operands are equal, false otherwise without sign
-  /// interpretation.
-  case equal
-  /// Yields `true` if the operands are unequal, false otherwise without sign
-  /// interpretation.
-  case notEqual
-
-  /// Interprets the operands as unsigned values and yields true if the first is
-  /// greater than the second.
-  case unsignedGreaterThan
-  /// Interprets the operands as unsigned values and yields true if the first is
-  /// greater than or equal to the second.
-  case unsignedGreaterThanOrEqual
-  /// Interprets the operands as unsigned values and yields true if the first is
-  /// less than the second.
-  case unsignedLessThan
-  /// Interprets the operands as unsigned values and yields true if the first is
-  /// less than or equal to the second.
-  case unsignedLessThanOrEqual
-
-  /// Interprets the operands as signed values and yields true if the first is
-  /// greater than the second.
-  case signedGreaterThan
-  /// Interprets the operands as signed values and yields true if the first is
-  /// greater than or equal to the second.
-  case signedGreaterThanOrEqual
-  /// Interprets the operands as signed values and yields true if the first is
-  /// less than the second.
-  case signedLessThan
-  /// Interprets the operands as signed values and yields true if the first is
-  /// less than or equal to the second.
-  case signedLessThanOrEqual
-
-  static let predicateMapping: [IntPredicate: LLVMIntPredicate] = [
-    .equal: LLVMIntEQ, .notEqual: LLVMIntNE, .unsignedGreaterThan: LLVMIntUGT,
-    .unsignedGreaterThanOrEqual: LLVMIntUGE, .unsignedLessThan: LLVMIntULT,
-    .unsignedLessThanOrEqual: LLVMIntULE, .signedGreaterThan: LLVMIntSGT,
-    .signedGreaterThanOrEqual: LLVMIntSGE, .signedLessThan: LLVMIntSLT,
-    .signedLessThanOrEqual: LLVMIntSLE,
-  ]
-
-  /// Retrieves the corresponding `LLVMIntPredicate`.
-  public var llvm: LLVMIntPredicate {
-    return IntPredicate.predicateMapping[self]!
-  }
-}
-
-/// The condition codes available for floating comparison instructions.
-public enum RealPredicate {
-  /// No comparison, always returns `false`.
-  case `false`
-  /// Ordered and equal.
-  case orderedEqual
-  /// Ordered greater than.
-  case orderedGreaterThan
-  /// Ordered greater than or equal.
-  case orderedGreaterThanOrEqual
-  /// Ordered less than.
-  case orderedLessThan
-  /// Ordered less than or equal.
-  case orderedLessThanOrEqual
-  /// Ordered and not equal.
-  case orderedNotEqual
-  /// Ordered (no nans).
-  case ordered
-  /// Unordered (either nans).
-  case unordered
-  /// Unordered or equal.
-  case unorderedEqual
-  /// Unordered or greater than.
-  case unorderedGreaterThan
-  /// Unordered or greater than or equal.
-  case unorderedGreaterThanOrEqual
-  /// Unordered or less than.
-  case unorderedLessThan
-  /// Unordered or less than or equal.
-  case unorderedLessThanOrEqual
-  /// Unordered or not equal.
-  case unorderedNotEqual
-  /// No comparison, always returns `true`.
-  case `true`
-
-  static let predicateMapping: [RealPredicate: LLVMRealPredicate] = [
-    .false: LLVMRealPredicateFalse, .orderedEqual: LLVMRealOEQ,
-    .orderedGreaterThan: LLVMRealOGT, .orderedGreaterThanOrEqual: LLVMRealOGE,
-    .orderedLessThan: LLVMRealOLT, .orderedLessThanOrEqual: LLVMRealOLE,
-    .orderedNotEqual: LLVMRealONE, .ordered: LLVMRealORD, .unordered: LLVMRealUNO,
-    .unorderedEqual: LLVMRealUEQ, .unorderedGreaterThan: LLVMRealUGT,
-    .unorderedGreaterThanOrEqual: LLVMRealUGE, .unorderedLessThan: LLVMRealULT,
-    .unorderedLessThanOrEqual: LLVMRealULE, .unorderedNotEqual: LLVMRealUNE,
-    .true: LLVMRealPredicateTrue,
-  ]
-
-  /// Retrieves the corresponding `LLVMRealPredicate`.
-  public var llvm: LLVMRealPredicate {
-    return RealPredicate.predicateMapping[self]!
-  }
-}
-
-/// `AtomicOrdering` enumerates available memory ordering semantics.
-///
-/// Atomic instructions (`cmpxchg`, `atomicrmw`, `fence`, `atomic load`, and
-/// `atomic store`) take ordering parameters that determine which other atomic
-/// instructions on the same address they synchronize with. These semantics are
-/// borrowed from Java and C++0x, but are somewhat more colloquial. If these
-/// descriptions aren’t precise enough, check those specs (see spec references
-/// in the atomics guide). fence instructions treat these orderings somewhat
-/// differently since they don’t take an address. See that instruction’s
-/// documentation for details.
-public enum AtomicOrdering: Comparable {
-  /// A load or store which is not atomic
-  case notAtomic
-  /// Lowest level of atomicity, guarantees somewhat sane results, lock free.
-  ///
-  /// The set of values that can be read is governed by the happens-before
-  /// partial order. A value cannot be read unless some operation wrote it. This
-  /// is intended to provide a guarantee strong enough to model Java’s
-  /// non-volatile shared variables. This ordering cannot be specified for
-  /// read-modify-write operations; it is not strong enough to make them atomic
-  /// in any interesting way.
-  case unordered
-  /// Guarantees that if you take all the operations affecting a specific
-  /// address, a consistent ordering exists.
-  ///
-  /// In addition to the guarantees of unordered, there is a single total order
-  /// for modifications by monotonic operations on each address. All
-  /// modification orders must be compatible with the happens-before order.
-  /// There is no guarantee that the modification orders can be combined to a
-  /// global total order for the whole program (and this often will not be
-  /// possible). The read in an atomic read-modify-write operation (cmpxchg and
-  /// atomicrmw) reads the value in the modification order immediately before
-  /// the value it writes. If one atomic read happens before another atomic read
-  /// of the same address, the later read must see the same value or a later
-  /// value in the address’s modification order. This disallows reordering of
-  /// monotonic (or stronger) operations on the same address. If an address is
-  /// written monotonic-ally by one thread, and other threads monotonic-ally
-  /// read that address repeatedly, the other threads must eventually see the
-  /// write. This corresponds to the C++0x/C1x memory_order_relaxed.
-  case monotonic
-  /// Acquire provides a barrier of the sort necessary to acquire a lock to
-  /// access other memory with normal loads and stores.
-  ///
-  /// In addition to the guarantees of monotonic, a synchronizes-with edge may
-  /// be formed with a release operation. This is intended to model C++’s
-  /// `memory_order_acquire`.
-  case acquire
-  /// Release is similar to Acquire, but with a barrier of the sort necessary to
-  /// release a lock.
-  ///
-  /// In addition to the guarantees of monotonic, if this operation writes a
-  /// value which is subsequently read by an acquire operation, it
-  /// synchronizes-with that operation. (This isn’t a complete description; see
-  /// the C++0x definition of a release sequence.) This corresponds to the
-  /// C++0x/C1x memory_order_release.
-  case release
-  /// provides both an Acquire and a Release barrier (for fences and operations
-  /// which both read and write memory).
-  ///
-  /// This corresponds to the C++0x/C1x memory_order_acq_rel.
-  case acquireRelease
-  /// Provides Acquire semantics for loads and Release semantics for stores.
-  ///
-  /// In addition to the guarantees of acq_rel (acquire for an operation that
-  /// only reads, release for an operation that only writes), there is a global
-  /// total order on all sequentially-consistent operations on all addresses,
-  /// which is consistent with the happens-before partial order and with the
-  /// modification orders of all the affected addresses. Each
-  /// sequentially-consistent read sees the last preceding write to the same
-  /// address in this global order. This corresponds to the C++0x/C1x
-  /// `memory_order_seq_cst` and Java volatile.
-  case sequentiallyConsistent
-
-  private static let orderingMapping: [AtomicOrdering: LLVMAtomicOrdering] = [
-    .notAtomic: LLVMAtomicOrderingNotAtomic,
-    .unordered: LLVMAtomicOrderingUnordered,
-    .monotonic: LLVMAtomicOrderingMonotonic,
-    .acquire: LLVMAtomicOrderingAcquire,
-    .release: LLVMAtomicOrderingRelease,
-    .acquireRelease: LLVMAtomicOrderingAcquireRelease,
-    .sequentiallyConsistent: LLVMAtomicOrderingSequentiallyConsistent,
-  ]
-
-  /// Returns whether the left atomic ordering is strictly weaker than the
-  /// right atomic order.
-  public static func <(lhs: AtomicOrdering, rhs: AtomicOrdering) -> Bool {
-    return lhs.llvm.rawValue < rhs.llvm.rawValue
-  }
-
-  /// Retrieves the corresponding `LLVMAtomicOrdering`.
-  public var llvm: LLVMAtomicOrdering {
-    return AtomicOrdering.orderingMapping[self]!
-  }
-}
-
-/// `AtomicReadModifyWriteOperation` enumerates the kinds of supported atomic
-/// read-write-modify operations.
-public enum AtomicReadModifyWriteOperation {
-  /// Set the new value and return the one old
-  ///
-  /// ```
-  /// *ptr = val
-  /// ```
-  case xchg
-  /// Add a value and return the old one
-  ///
-  /// ```
-  /// *ptr = *ptr + val
-  /// ```
-  case add
-  /// Subtract a value and return the old one
-  ///
-  /// ```
-  /// *ptr = *ptr - val
-  /// ```
-  case sub
-  /// And a value and return the old one
-  ///
-  /// ```
-  /// *ptr = *ptr & val
-  /// ```
-  case and
-  /// Not-And a value and return the old one
-  ///
-  /// ```
-  /// *ptr = ~(*ptr & val)
-  /// ```
-  case nand
-  /// OR a value and return the old one
-  ///
-  /// ```
-  /// *ptr = *ptr | val
-  /// ```
-  case or
-  /// Xor a value and return the old one
-  ///
-  /// ```
-  /// *ptr = *ptr ^ val
-  /// ```
-  case xor
-  /// Sets the value if it's greater than the original using a signed comparison
-  /// and return the old one.
-  ///
-  /// ```
-  /// // Using a signed comparison
-  /// *ptr = *ptr > val ? *ptr : val
-  /// ```
-  case max
-  /// Sets the value if it's Smaller than the original using a signed comparison
-  /// and return the old one.
-  ///
-  /// ```
-  /// // Using a signed comparison
-  /// *ptr = *ptr < val ? *ptr : val
-  /// ```
-  case min
-  /// Sets the value if it's greater than the original using an unsigned
-  /// comparison and return the old one.
-  ///
-  /// ```
-  /// // Using an unsigned comparison
-  /// *ptr = *ptr > val ? *ptr : val
-  /// ```
-  case umax
-  /// Sets the value if it's greater than the original using an unsigned
-  /// comparison and return the old one.
-  ///
-  /// ```
-  /// // Using an unsigned comparison
-  /// *ptr = *ptr < val ? *ptr : val
-  /// ```
-  case umin
-
-  static let atomicRMWMapping: [AtomicReadModifyWriteOperation: LLVMAtomicRMWBinOp] = [
-    .xchg: LLVMAtomicRMWBinOpXchg, .add: LLVMAtomicRMWBinOpAdd,
-    .sub: LLVMAtomicRMWBinOpSub, .and: LLVMAtomicRMWBinOpAnd,
-    .nand: LLVMAtomicRMWBinOpNand, .or: LLVMAtomicRMWBinOpOr,
-    .xor: LLVMAtomicRMWBinOpXor, .max: LLVMAtomicRMWBinOpMax,
-    .min: LLVMAtomicRMWBinOpMin, .umax: LLVMAtomicRMWBinOpUMax,
-    .umin: LLVMAtomicRMWBinOpUMin,
-  ]
-
-  /// Retrieves the corresponding `LLVMAtomicRMWBinOp`.
-  public var llvm: LLVMAtomicRMWBinOp {
-    return AtomicReadModifyWriteOperation.atomicRMWMapping[self]!
-  }
-}
-
-extension Module {
-  /// Searches for and retrieves a global variable with the given name in this
-  /// module if that name references an existing global variable.
-  ///
-  /// - parameter name: The name of the global to reference.
-  ///
-  /// - returns: A value representing the referenced global if it exists.
-  public func global(named name: String) -> Global? {
-    guard let ref = LLVMGetNamedGlobal(llvm, name) else { return nil }
-    return Global(llvm: ref)
-  }
-
-  /// Searches for and retrieves a type with the given name in this module if
-  /// that name references an existing type.
-  ///
-  /// - parameter name: The name of the type to create.
-  ///
-  /// - returns: A representation of the newly created type with the given name
-  ///   or nil if such a representation could not be created.
-  public func type(named name: String) -> IRType? {
-    guard let type = LLVMGetTypeByName(llvm, name) else { return nil }
-    return convertType(type)
-  }
-
-  /// Searches for and retrieves a function with the given name in this module
-  /// if that name references an existing function.
-  ///
-  /// - parameter name: The name of the function to create.
-  ///
-  /// - returns: A representation of the newly created function with the given
-  /// name or nil if such a representation could not be created.
-  public func function(named name: String) -> Function? {
-    guard let fn = LLVMGetNamedFunction(llvm, name) else { return nil }
-    return Function(llvm: fn)
-  }
-
-  /// Searches for and retrieves a comdat section with the given name in this
-  /// module.  If none is found, one with that name is created and returned.
-  ///
-  /// - parameter name: The name of the comdat section to create.
-  ///
-  /// - returns: A representation of the newly created comdat section with the
-  ///   given name.
-  public func comdat(named name: String) -> Comdat {
-    guard let comdat = LLVMGetOrInsertComdat(llvm, name) else { fatalError() }
-    return Comdat(llvm: comdat)
-  }
-
-  /// Build a named global of the given type.
-  ///
-  /// - parameter name: The name of the newly inserted global value.
-  /// - parameter type: The type of the newly inserted global value.
-  /// - parameter addressSpace: The optional address space where the global
-  ///   variable resides.
-  ///
-  /// - returns: A value representing the newly inserted global variable.
-  public func addGlobal(_ name: String, type: IRType, addressSpace: Int? = nil) -> Global {
-    let val: LLVMValueRef
-    if let addressSpace = addressSpace {
-      val = LLVMAddGlobalInAddressSpace(llvm, type.asLLVM(), name, UInt32(addressSpace))
-    } else {
-      val = LLVMAddGlobal(llvm, type.asLLVM(), name)
-    }
-    return Global(llvm: val)
-  }
-
-  /// Build a named global of the given type.
-  ///
-  /// - parameter name: The name of the newly inserted global value.
-  /// - parameter initializer: The initial value for the global variable.
-  /// - parameter addressSpace: The optional address space where the global
-  ///   variable resides.
-  ///
-  /// - returns: A value representing the newly inserted global variable.
-  public func addGlobal(_ name: String, initializer: IRValue, addressSpace: Int? = nil) -> Global {
-    let global = addGlobal(name, type: initializer.type)
-    global.initializer = initializer
-    return global
-  }
-
-  /// Build a named global string consisting of an array of `i8` type filled in
-  /// with the nul terminated string value.
-  ///
-  /// - parameter name: The name of the newly inserted global string value.
-  /// - parameter value: The character contents of the newly inserted global.
-  ///
-  /// - returns: A value representing the newly inserted global string variable.
-  public func addGlobalString(name: String, value: String) -> Global {
-    let length = value.utf8.count
-
-    var global = addGlobal(name, type:
-      ArrayType(elementType: IntType.int8, count: length + 1))
-
-    global.alignment = Alignment(1)
-    global.initializer = value
-
-    return global
-  }
-
-  /// Build a named alias to a global value or a constant expression.
-  ///
-  /// Aliases, unlike function or variables, don’t create any new data. They are
-  /// just a new symbol and metadata for an existing position.
-  ///
-  /// - parameter name: The name of the newly inserted alias.
-  /// - parameter aliasee: The value or constant to alias.
-  /// - parameter type: The type of the aliased value or expression.
-  ///
-  /// - returns: A value representing the newly created alias.
-  public func addAlias(name: String, to aliasee: IRGlobal, type: IRType) -> Alias {
-    return Alias(llvm: LLVMAddAlias(llvm, type.asLLVM(), aliasee.asLLVM(), name))
-  }
-}
-
 /// An `IRBuilder` is a helper object that generates LLVM instructions.  IR
 /// Builders keep track of a position within a function or basic block and has
 /// methods to insert instructions at that position.
@@ -436,8 +19,14 @@ public class IRBuilder {
     self.llvm = LLVMCreateBuilderInContext(module.context.llvm)
   }
 
-  // MARK: IR Navigation
+  deinit {
+    LLVMDisposeBuilder(llvm)
+  }
+}
 
+// MARK: IR Navigation
+
+extension IRBuilder {
   /// Repositions the IR Builder at the end of the given basic block.
   ///
   /// - parameter block: The basic block to reposition the IR Builder after.
@@ -471,8 +60,6 @@ public class IRBuilder {
     LLVMClearInsertionPosition(llvm)
   }
 
-  // MARK: Instruction Insertion
-
   /// Gets the basic block built instructions will be inserted into.
   public var insertBlock: BasicBlock? {
     guard let blockRef = LLVMGetInsertBlock(llvm) else { return nil }
@@ -495,9 +82,11 @@ public class IRBuilder {
       LLVMInsertIntoBuilder(llvm, inst.asLLVM())
     }
   }
+}
 
-  // MARK: Convenience Instructions
+// MARK: Convenience Instructions
 
+extension IRBuilder {
   /// Builds the specified binary operation instruction with the given arguments.
   ///
   /// - parameter op: The operation to build.
@@ -524,9 +113,11 @@ public class IRBuilder {
   public func buildCast(_ op: OpCode.Cast, value: IRValue, type: IRType, name: String = "") -> IRValue {
     return LLVMBuildCast(llvm, op.llvm, value.asLLVM(), type.asLLVM(), name)
   }
+}
 
-  // MARK: Arithmetic Instructions
+// MARK: Arithmetic Instructions
 
+extension IRBuilder {
   /// Build a negation instruction with the given value as an operand.
   ///
   /// Whether an integer or floating point negate instruction is built is
@@ -790,9 +381,11 @@ public class IRBuilder {
     }
     return LLVMBuildFCmp(llvm, predicate.llvm, lhsVal, rhsVal, name)
   }
+}
 
-  // MARK: Logical Instructions
+// MARK: Logical Instructions
 
+extension IRBuilder {
   /// Build a bitwise logical not with the given value as an operand.
   ///
   /// - parameter val: The value to negate.
@@ -875,9 +468,11 @@ public class IRBuilder {
       return LLVMBuildLShr(llvm, lhs.asLLVM(), rhs.asLLVM(), name)
     }
   }
+}
 
-  // MARK: Conditional Instructions
+// MARK: Conditional Instructions
 
+extension IRBuilder {
   /// Build a phi node with the given type acting as the result of any incoming
   /// basic blocks.
   ///
@@ -940,9 +535,11 @@ public class IRBuilder {
                                         `else`.asLLVM(),
                                         UInt32(caseCount))!)
   }
+}
 
-  // MARK: Declaration Instructions
+// MARK: Declaration Instructions
 
+extension IRBuilder {
   /// Build a named function body with the given type.
   ///
   /// - parameter name: The name of the newly defined function.
@@ -969,9 +566,11 @@ public class IRBuilder {
     }
     return type
   }
+}
 
-  // MARK: Terminator Instructions
+// MARK: Terminator Instructions
 
+extension IRBuilder {
   /// Build an unconditional branch to the given basic block.
   ///
   /// The `br` instruction is used to cause control flow to transfer to a
@@ -1110,9 +709,11 @@ public class IRBuilder {
       return Call(llvm: LLVMBuildCall(llvm, fn.asLLVM(), buf.baseAddress!, UInt32(buf.count), name))
     }
   }
+}
 
-  // MARK: Exception Handling Instructions
+// MARK: Exception Handling Instructions
 
+extension IRBuilder {
   /// Build a call to the given function with the given arguments with the
   /// possibility of control transfering to either the `next` basic block or
   /// the `catch` basic block if an exception occurs.
@@ -1222,9 +823,11 @@ public class IRBuilder {
   public func buildVAArg(_ list: IRValue, type: IRType, name: String = "") -> IRValue {
     return LLVMBuildVAArg(llvm, list.asLLVM(), type.asLLVM(), name)
   }
+}
 
-  // MARK: Memory Access Instructions
+// MARK: Memory Access Instructions
 
+extension IRBuilder {
   /// Build an `alloca` to allocate stack memory to hold a value of the given
   /// type.
   ///
@@ -1370,9 +973,11 @@ public class IRBuilder {
                                 name: String = "") -> IRValue {
     return LLVMBuildExtractValue(llvm, value.asLLVM(), UInt32(index), name)
   }
+}
 
-  // MARK: Null Test Instructions
+// MARK: Null Test Instructions
 
+extension IRBuilder {
   /// Build a comparision instruction that returns whether the given operand is
   /// `null`.
   ///
@@ -1396,9 +1001,11 @@ public class IRBuilder {
   public func buildIsNotNull(_ val: IRValue, name: String = "") -> IRValue {
     return LLVMBuildIsNotNull(llvm, val.asLLVM(), name)
   }
+}
 
-  // MARK: Conversion Instructions
+// MARK: Conversion Instructions
 
+extension IRBuilder {
   /// Build an instruction that either performs a truncation or a bitcast of
   /// the given value to a value of the given type.
   ///
@@ -1632,9 +1239,11 @@ public class IRBuilder {
     )
     return LLVMBuildPtrDiff(llvm, lhs.asLLVM(), rhs.asLLVM(), name)
   }
+}
 
-  // MARK: Atomic Instructions
+// MARK: Atomic Instructions
 
+extension IRBuilder {
   /// Build a fence instruction that introduces "happens-before" edges between
   /// operations.
   ///
@@ -1734,9 +1343,11 @@ public class IRBuilder {
   ) -> IRValue {
     return LLVMBuildAtomicRMW(llvm, atomicOp.llvm, ptr.asLLVM(), value.asLLVM(), ordering.llvm, singleThreaded.llvm)
   }
+}
 
-  // MARK: C Standard Library Instructions
+// MARK: C Standard Library Instructions
 
+extension IRBuilder {
   /// Build a call to the C standard library `malloc` instruction.
   /// ```
   /// (type *)malloc(sizeof(type));
@@ -1769,9 +1380,11 @@ public class IRBuilder {
   public func buildFree(_ ptr: IRValue) -> IRValue {
     return LLVMBuildFree(llvm, ptr.asLLVM())
   }
+}
 
-  // MARK: Aggregate Instructions
+// MARK: Aggregate Instructions
 
+extension IRBuilder {
   /// Build an instruction to insert a value into a member field in an
   /// aggregate value.
   ///
@@ -1802,9 +1415,11 @@ public class IRBuilder {
   func buildExtractValue(aggregate: IRValue, index: Int, name: String = "") -> IRValue {
     return LLVMBuildExtractValue(llvm, aggregate.asLLVM(), UInt32(index), name)
   }
+}
 
-  // MARK: Vector Instructions
+// MARK: Vector Instructions
 
+extension IRBuilder {
   /// Build a vector insert instruction to nondestructively insert the given
   /// value into the given vector.
   ///
@@ -1856,9 +1471,11 @@ public class IRBuilder {
     }
     return LLVMBuildShuffleVector(llvm, vector1.asLLVM(), vector2.asLLVM(), mask.asLLVM(), name)
   }
+}
 
-  // MARK: Global Variable Instructions
+// MARK: Global Variable Instructions
 
+extension IRBuilder {
   /// Build a named global of the given type.
   ///
   /// - parameter name: The name of the newly inserted global value.
@@ -1931,9 +1548,11 @@ public class IRBuilder {
   public func addAlias(name: String, to aliasee: IRGlobal, type: IRType) -> Alias {
     return self.module.addAlias(name: name, to: aliasee, type: type)
   }
+}
 
-  // MARK: Inline Assembly
+// MARK: Inline Assembly
 
+extension IRBuilder {
   /// Build a value representing an inline assembly expression (as opposed to
   /// module-level inline assembly).
   ///
@@ -1972,10 +1591,6 @@ public class IRBuilder {
   ///   expression.
   public func buildInlineAssembly(_ asm: String, type: FunctionType, constraints: String = "", hasSideEffects: Bool = true, needsAlignedStack: Bool = true) -> IRValue {
     return LLVMConstInlineAsm(type.asLLVM(), asm, constraints, hasSideEffects.llvm, needsAlignedStack.llvm)
-  }
-
-  deinit {
-    LLVMDisposeBuilder(llvm)
   }
 }
